+++ e2e-tests/tests/common/mod.rs
@@ -1,0 +1,65 @@
+pub use candid::utils::{decode_args, encode_args, ArgumentDecoder, ArgumentEncoder};
+pub use ic_state_machine_tests::{CanisterId, ErrorCode, StateMachine, UserError, WasmResult};
+
+#[derive(Debug)]
+pub enum CallError {
+    Reject(String),
+    UserError(UserError),
+}
+
+/// A helper function that we use to implement both [`call_candid`] and
+/// [`query_candid`].
+fn with_candid<Input, Output>(
+    input: Input,
+    f: impl FnOnce(Vec<u8>) -> Result<WasmResult, UserError>,
+) -> Result<Output, CallError>
+where
+    Input: ArgumentEncoder,
+    Output: for<'a> ArgumentDecoder<'a>,
+{
+    let in_bytes = encode_args(input).expect("failed to encode args");
+    match f(in_bytes) {
+        Ok(WasmResult::Reply(out_bytes)) => Ok(decode_args(&out_bytes).unwrap_or_else(|e| {
+            panic!(
+                "Failed to decode bytes {:?} as candid type: {}",
+                std::any::type_name::<Output>(),
+                e
+            )
+        })),
+        Ok(WasmResult::Reject(message)) => Err(CallError::Reject(message)),
+        Err(user_error) => Err(CallError::UserError(user_error)),
+    }
+}
+
+/// Call a canister candid method.
+pub fn call_candid<Input, Output>(
+    env: &StateMachine,
+    canister_id: CanisterId,
+    method: &str,
+    input: Input,
+) -> Result<Output, CallError>
+where
+    Input: ArgumentEncoder,
+    Output: for<'a> ArgumentDecoder<'a>,
+{
+    with_candid(input, |bytes| {
+        env.execute_ingress(canister_id, method, bytes)
+    })
+}
+
+/// Query a canister candid method.
+// This function is indeed used by tests.
+// Might be a bug with cargo.
+#[allow(dead_code)] 
+pub fn query_candid<Input, Output>(
+    env: &StateMachine,
+    canister_id: CanisterId,
+    method: &str,
+    input: Input,
+) -> Result<Output, CallError>
+where
+    Input: ArgumentEncoder,
+    Output: for<'a> ArgumentDecoder<'a>,
+{
+    with_candid(input, |bytes| env.query(canister_id, method, bytes))
+}
