+++ e2e-tests/tests/e2e.rs
@@ -1,67 +1,8 @@
-use candid::utils::{decode_args, encode_args, ArgumentDecoder, ArgumentEncoder};
 use ic_cdk_e2e_tests::cargo_build_canister;
-use ic_state_machine_tests::{CanisterId, ErrorCode, StateMachine, UserError, WasmResult};
 use serde_bytes::ByteBuf;
 
-#[derive(Debug)]
-enum CallError {
-    Reject(String),
-    UserError(UserError),
-}
-
-/// A helper function that we use to implement both [`call_candid`] and
-/// [`query_candid`].
-fn with_candid<Input, Output>(
-    input: Input,
-    f: impl FnOnce(Vec<u8>) -> Result<WasmResult, UserError>,
-) -> Result<Output, CallError>
-where
-    Input: ArgumentEncoder,
-    Output: for<'a> ArgumentDecoder<'a>,
-{
-    let in_bytes = encode_args(input).expect("failed to encode args");
-    match f(in_bytes) {
-        Ok(WasmResult::Reply(out_bytes)) => Ok(decode_args(&out_bytes).unwrap_or_else(|e| {
-            panic!(
-                "Failed to decode bytes {:?} as candid type: {}",
-                std::any::type_name::<Output>(),
-                e
-            )
-        })),
-        Ok(WasmResult::Reject(message)) => Err(CallError::Reject(message)),
-        Err(user_error) => Err(CallError::UserError(user_error)),
-    }
-}
-
-/// Call a canister candid method.
-fn call_candid<Input, Output>(
-    env: &StateMachine,
-    canister_id: CanisterId,
-    method: &str,
-    input: Input,
-) -> Result<Output, CallError>
-where
-    Input: ArgumentEncoder,
-    Output: for<'a> ArgumentDecoder<'a>,
-{
-    with_candid(input, |bytes| {
-        env.execute_ingress(canister_id, method, bytes)
-    })
-}
-
-/// Query a canister candid method.
-fn query_candid<Input, Output>(
-    env: &StateMachine,
-    canister_id: CanisterId,
-    method: &str,
-    input: Input,
-) -> Result<Output, CallError>
-where
-    Input: ArgumentEncoder,
-    Output: for<'a> ArgumentDecoder<'a>,
-{
-    with_candid(input, |bytes| env.query(canister_id, method, bytes))
-}
+mod common;
+use common::*;
 
 /// Checks that a canister that uses [`ic_cdk::storage::stable_store`]
 /// and [`ic_cdk::storage::stable_restore`] functions can keep its data
@@ -168,16 +109,3 @@ fn test_api_call() {
     assert!(result > 0);
 }
 
-#[test]
-fn test_management_canister() {
-    let env = StateMachine::new();
-    let rev = cargo_build_canister("call-management");
-    let canister_id = env.install_canister(rev, vec![], None).unwrap();
-
-    let (_result,): (CanisterId,) =
-        call_candid(&env, canister_id, "call_create_canister", ()).expect("failed to call call_create_canister");
-
-    let (result,): (Vec<u8>,) =
-        call_candid(&env, canister_id, "call_raw_rand", ()).expect("failed to call call_raw_rand");
-    assert_eq!(result.len(), 32);
-}
